cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commerce-logs-cli"
version = "0.1.0"
description = "CLI tools for Commerce Logs Pipeline administration and troubleshooting"
authors = [{name = "Commerce Logs Pipeline Team"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "azure-storage-blob>=12.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
list-blobs = "commerce_logs_cli.list_blobs_cli:main"
setup-config = "commerce_logs_cli.setup_config_cli:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["commerce_logs_cli"] 